change price type from long to integer

+++ src/main/java/io/chainmind/myriad/domain/dto/distribution/DistributionListItem.java
@@ -34,8 +34,8 @@ public class DistributionListItem {
 	// gift voucher
 	private String productName;
 	private String productCode;
-	private Long productMarketPrice;
-	private Long productExchangePrice;
+	private Integer productMarketPrice;
+	private Integer productExchangePrice;
 
 
 }
